+++ lib/osm/member.rb
@@ -74,6 +74,8 @@ module Osm
     #   @return [Fixnum] the grouping within the section that the member belongs to
     # @!attribute [rw] grouping_leader
     #   @return [Fixnum] whether the member is the grouping leader (0=no, 1=seconder/APL, 2=sixer/PL)
+    # @!attribute [rw] grouping_name
+    #   @return [Fixnum] the grouping within the section that the member belongs to
     # @!attribute [rw] joined
     #   @return [Date] when the member joined the section
     # @!attribute [rw] age
@@ -116,6 +118,7 @@ module Osm
     attribute :custom8, :type => String, :default => ''
     attribute :custom9, :type => String, :default => ''
     attribute :grouping_id, :type => Integer
+    attribute :grouping_name, :type => String, :default => ''
     attribute :grouping_leader, :type => Integer
     attribute :joined, :type => Date
     attribute :age, :type => String
@@ -125,7 +128,7 @@ module Osm
                     :phone1, :phone2, :phone3, :phone4, :address, :address2, :date_of_birth, :started,
                     :joining_in_years, :parents, :notes, :medical, :religion, :school, :ethnicity, :subs,
                     :custom1, :custom2, :custom3, :custom4, :custom5, :custom6, :custom7, :custom8, :custom9,
-                    :grouping_id, :grouping_leader, :joined, :age, :joined_years
+                    :grouping_id, :grouping_name, :grouping_leader, :joined, :age, :joined_years
 
     validates_numericality_of :id, :only_integer=>true, :greater_than=>0, :unless => Proc.new { |r| r.id.nil? }
     validates_numericality_of :section_id, :only_integer=>true, :greater_than=>0
@@ -197,6 +200,7 @@ module Osm
           :custom8 => item['custom8'],
           :custom9 => item['custom9'],
           :grouping_id => Osm::to_i_or_nil(item['patrolid']),
+          :grouping_name => Osm::to_i_or_nil(item['patrol']),
           :grouping_leader => Osm::to_i_or_nil(item['patrolleader']),
           :joined => Osm::parse_date(item['joined']),
           :age => item['age'].gsub(' ', ''),